+++ src/vec_utils.c
@@ -1,18 +1,18 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   vec_get.c                                          :+:      :+:    :+:   */
+/*   vec_utils.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: awillems <awillems@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/24 12:46:44 by awillems          #+#    #+#             */
-/*   Updated: 2022/05/24 12:49:35 by awillems         ###   ########.fr       */
+/*   Updated: 2022/05/30 15:45:03 by awillems         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "vector_template.h"
 
-void	*vec_get(t_vec *vec, int index)
+void	*vec_get(t_vec *vec, size_t index)
 {
 	return (vec->buffer + (index * vec->size));
 }
